requirements.txt
Flask==2.3.2
gunicorn==20.1.0
aiohttp==3.8.4
pymongo[srv]==3.12.3
python-telegram-bot==13.13
telethon
pytube
